// Desarrollo_Web_en_entorno_cliente/Casa/Trabajos/Unit1 - JS/foodscore-client/src/interfaces/user.ts
export interface User {
    id?: number;
    name: string;
    email: string;
    password?: string;
    avatar: string;
    lat: number;
    lng: number;
    me?: boolean;
    token?:string;
}

export interface UserLogin {
    email: string;
    password: string;
    lat?: number;
    lng?: number;
}